Fix gauntlet findings: path traversal, version sync, dedup assertions
Critical:
- install.py: replace string startswith() with Path.is_relative_to()
  for path containment check (fixes traversal bypass)

Major:
- Sync __init__.py version to 0.1.1 (was 0.1.0, pyproject.toml was 0.1.1)
- Extract shared assertion logic to brunnr/assertions.py, used by both
  brunnr/commands/eval.py and tests/harness.py (eliminates 150+ lines
  of duplication)
- Replace stub path traversal tests with real integration tests that
  exercise the actual sanitization logic (6 tests, 4 new)

78 tests pass.
