+++ brunnr/__init__.py
@@ -1,3 +1,3 @@
 """brunnr — Security scanner and skill registry for agent tool descriptions."""
 
-__version__ = "0.1.0"
+__version__ = "0.1.1"
