+++ tests/harness.py
@@ -14,35 +14,30 @@ Usage:
 
 import argparse
 import json
-import os
-import re
 import sys
 import time
 from pathlib import Path
 from typing import Any
 
+# Ensure brunnr package is importable when run as standalone script
 REPO_ROOT = Path(__file__).resolve().parent.parent
+sys.path.insert(0, str(REPO_ROOT))
+
+from brunnr.assertions import (
+    parse_frontmatter,
+    extract_score,
+    extract_criteria,
+    check_has_rewrite,
+    run_assertion,
+)
+
 SKILLS_DIR = REPO_ROOT / "skills"
 TESTS_DIR = REPO_ROOT / "tests"
 
-# ---------------------------------------------------------------------------
-# Frontmatter parsing (lightweight, no pyyaml dependency for the basic case)
-# ---------------------------------------------------------------------------
 
-def parse_frontmatter(text: str) -> dict[str, str]:
-    """Extract YAML frontmatter from a SKILL.md file."""
-    if not text.startswith("---"):
-        return {}
-    parts = text.split("---", 2)
-    if len(parts) < 3:
-        return {}
-    fm = {}
-    for line in parts[1].strip().splitlines():
-        if ":" in line:
-            key, val = line.split(":", 1)
-            fm[key.strip()] = val.strip().strip('"').strip("'")
-    return fm
-
+# ---------------------------------------------------------------------------
+# Data loading
+# ---------------------------------------------------------------------------
 
 def load_skill_prompt(skill_slug: str) -> str:
     """Read the full SKILL.md file as the system prompt for the skill."""
@@ -98,173 +93,6 @@ def call_claude(
     )
 
     return response.content[0].text
-
-
-# ---------------------------------------------------------------------------
-# Assertion engine
-# ---------------------------------------------------------------------------
-
-def extract_score(text: str) -> int | None:
-    """Extract the primary score (N/5) from skill output."""
-    # Match "Score: N/5" or "**Score: N/5**" or "Score: N / 5"
-    match = re.search(r"[Ss]core:\s*\**\s*(\d)\s*/\s*5", text)
-    if match:
-        return int(match.group(1))
-    return None
-
-
-def extract_criteria_results(text: str) -> dict[str, str]:
-    """Extract per-criterion pass/fail from the output table."""
-    results = {}
-    # Match table rows: | Criterion | pass/fail | notes |
-    for match in re.finditer(
-        r"\|\s*([\w\s]+?)\s*\|\s*(pass|fail)\s*\|",
-        text,
-        re.IGNORECASE,
-    ):
-        criterion = match.group(1).strip().lower()
-        result = match.group(2).strip().lower()
-        results[criterion] = result
-    return results
-
-
-def check_has_rewrite(text: str) -> bool:
-    """Check if the output contains a rewrite section."""
-    return bool(
-        re.search(r"[Rr]ewr(itten|ite)", text)
-        and re.search(r"[45]/5", text[text.lower().find("rewrit"):] if "rewrit" in text.lower() else text)
-    )
-
-
-def run_assertion(assertion: dict[str, Any], output: str) -> dict[str, Any]:
-    """
-    Run a single assertion against the output.
-
-    Assertion types:
-      - score_range: { "min": 0, "max": 2 } — extracted score must be in range
-      - score_exact: 5 — extracted score must match exactly
-      - has_rewrite: true — output must contain a rewrite section
-      - has_table: true — output must contain a criterion table
-      - contains: "string" — output must contain the string (case-insensitive)
-      - criteria_fail: ["output shape", "cost signal"] — these criteria must show "fail"
-      - criteria_pass: ["trigger clarity"] — these criteria must show "pass"
-      - format_valid: true — output must follow the expected markdown format
-    """
-    atype = assertion["type"]
-    expected = assertion["expected"]
-    result = {"assertion": assertion, "passed": False, "detail": ""}
-
-    if atype == "score_range":
-        score = extract_score(output)
-        if score is None:
-            result["detail"] = "Could not extract score from output"
-        elif expected["min"] <= score <= expected["max"]:
-            result["passed"] = True
-            result["detail"] = f"Score {score} is in range [{expected['min']}, {expected['max']}]"
-        else:
-            result["detail"] = f"Score {score} is NOT in range [{expected['min']}, {expected['max']}]"
-
-    elif atype == "score_exact":
-        score = extract_score(output)
-        if score is None:
-            result["detail"] = "Could not extract score from output"
-        elif score == expected:
-            result["passed"] = True
-            result["detail"] = f"Score {score} matches expected {expected}"
-        else:
-            result["detail"] = f"Score {score} does not match expected {expected}"
-
-    elif atype == "has_rewrite":
-        has_it = check_has_rewrite(output)
-        if has_it == expected:
-            result["passed"] = True
-            result["detail"] = f"Rewrite section {'found' if has_it else 'not found'} as expected"
-        else:
-            result["detail"] = f"Rewrite section {'found' if has_it else 'not found'}, expected {'present' if expected else 'absent'}"
-
-    elif atype == "has_table":
-        has_table = bool(re.search(r"\|.*\|.*\|.*\|", output))
-        if has_table == expected:
-            result["passed"] = True
-            result["detail"] = f"Criteria table {'found' if has_table else 'not found'} as expected"
-        else:
-            result["detail"] = f"Criteria table {'found' if has_table else 'not found'}, expected {'present' if expected else 'absent'}"
-
-    elif atype == "contains":
-        found = expected.lower() in output.lower()
-        if found:
-            result["passed"] = True
-            result["detail"] = f"Output contains '{expected}'"
-        else:
-            result["detail"] = f"Output does NOT contain '{expected}'"
-
-    elif atype == "criteria_fail":
-        criteria = extract_criteria_results(output)
-        all_failed = True
-        missing = []
-        for crit in expected:
-            crit_lower = crit.lower()
-            matched = False
-            for key, val in criteria.items():
-                if crit_lower in key:
-                    if val != "fail":
-                        all_failed = False
-                        missing.append(f"{crit} is '{val}', expected 'fail'")
-                    matched = True
-                    break
-            if not matched:
-                all_failed = False
-                missing.append(f"{crit} not found in output table")
-        if all_failed:
-            result["passed"] = True
-            result["detail"] = f"All expected criteria marked fail: {expected}"
-        else:
-            result["detail"] = f"Criteria mismatch: {'; '.join(missing)}"
-
-    elif atype == "criteria_pass":
-        criteria = extract_criteria_results(output)
-        all_passed = True
-        missing = []
-        for crit in expected:
-            crit_lower = crit.lower()
-            matched = False
-            for key, val in criteria.items():
-                if crit_lower in key:
-                    if val != "pass":
-                        all_passed = False
-                        missing.append(f"{crit} is '{val}', expected 'pass'")
-                    matched = True
-                    break
-            if not matched:
-                all_passed = False
-                missing.append(f"{crit} not found in output table")
-        if all_passed:
-            result["passed"] = True
-            result["detail"] = f"All expected criteria marked pass: {expected}"
-        else:
-            result["detail"] = f"Criteria mismatch: {'; '.join(missing)}"
-
-    elif atype == "format_valid":
-        # Check for the core structure: tool name header, score line, table, rewrite
-        has_header = bool(re.search(r"^###?\s+", output, re.MULTILINE))
-        has_score = extract_score(output) is not None
-        has_table = bool(re.search(r"\|.*[Cc]riterion.*\|", output))
-        checks = {
-            "has_header": has_header,
-            "has_score_line": has_score,
-            "has_criteria_table": has_table,
-        }
-        if all(checks.values()):
-            result["passed"] = True
-            result["detail"] = "Output format valid: header, score, table all present"
-        else:
-            failed = [k for k, v in checks.items() if not v]
-            result["detail"] = f"Format invalid — missing: {', '.join(failed)}"
-
-    else:
-        result["detail"] = f"Unknown assertion type: {atype}"
-
-    return result
 
 
 # ---------------------------------------------------------------------------
@@ -327,7 +155,7 @@ def run_test_case(
     return {
         "case_id": case_id,
         "description": description,
-        "output": output[:2000],  # Truncate for storage
+        "output": output[:2000],
         "assertions": assertion_results,
         "duration_s": duration,
         "all_passed": all(r["passed"] for r in assertion_results),
